+++ knishioclient/__init__.py
@@ -17,7 +17,7 @@ from .storage import (
     create_default_secret_storage,
 )
 
-__version__ = '1.2.0'
+__version__ = '1.2.1'
 
 name = "knishioclient"